Add -mc option in faust2w32max6 and faust2w64max6.

+++ architecture/max-msp/max-msp64.cpp
@@ -138,6 +138,7 @@ using namespace std;
 #include "z_dsp.h"
 #include "jpatcher_api.h"
 #include <string.h>
+#include <stdint.h>
 
 #define ASSIST_INLET 	1
 #define ASSIST_OUTLET 	2
@@ -713,10 +714,10 @@ void faust_perform64(t_faust* x, t_object* dsp64, double** ins, long numins, dou
 
 /*--------------------------------------------------------------------------*/
 void faust_dsp64(t_faust* x, t_object* dsp64, short* count, double samplerate, long maxvectorsize, long flags)
-{ 
+{
 #ifdef MC_VERSION
     // We need to know the real number of inputs to adapt faust_perform64
-    long inputs = (long)object_method(dsp64, gensym("getnuminputchannels"), x, 0);
+    intptr_t inputs = (intptr_t)object_method(dsp64, gensym("getnuminputchannels"), x, 0);
     delete x->m_mc_dsp;
     // x->m_mc_dsp will not delete the x->m_dsp object
     x->m_mc_dsp = new dsp_adapter(x->m_dsp, inputs, x->m_Outputs, maxvectorsize, false);
